chore: update files of tooltip-handler
