--- packages/vchart/src/plugin/components/tooltip-handler/canvas/index.ts
@@ -1,1 +1,0 @@
-export * from './canvas-tooltip-handler';
