--- packages/vchart/src/plugin/components/tooltip-handler/dom/constant.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_TOOLTIP_Z_INDEX = '99999999999999';
