--- packages/vchart/src/plugin/components/tooltip-handler/dom/index.ts
@@ -1,1 +1,0 @@
-export * from './dom-tooltip-handler';
