--- packages/vchart/src/plugin/components/tooltip-handler/dom/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './common';
-export * from './style';
