+++ packages/vchart/src/plugin/components/tooltip-handler/index.ts
@@ -1,2 +1,2 @@
-export * from './dom';
-export * from './canvas';
+export * from './dom-tooltip-handler';
+export * from './canvas-tooltip-handler';
