--- packages/vchart/src/plugin/components/tooltip-handler/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './common';
-export * from './position';
-export * from './attribute';
